refactor: Move tasks to root namespace
