+++ examples/Greeter/GreetWith3Fails.php
@@ -3,7 +3,7 @@
 namespace Examples\Greeter;
 
 use Generator;
-use Tsqm\Tasks\Task;
+use Tsqm\Task;
 
 class GreetWith3Fails
 {
