+++ examples/Greeter/GreetWith3PurchaseFailsAnd3Retries.php
@@ -3,8 +3,8 @@
 namespace Examples\Greeter;
 
 use Generator;
-use Tsqm\Tasks\RetryPolicy;
-use Tsqm\Tasks\Task;
+use Tsqm\RetryPolicy;
+use Tsqm\Task;
 
 class GreetWith3PurchaseFailsAnd3Retries
 {
